refactor(grading): one transcript_rules evaluator in core, taking the model (#685 stage 4)
evaluate_transcript_rules moves from runner/grading.py to
core/grading/transcript.py and takes the validated TranscriptRulesConfig
instead of a dump. TranscriptChecker is deleted; the core engine's own fold
moves to core/grading/transcript_buckets.py unchanged, so no verdict moves on
either substrate in this commit.

The runner's call site in service.py is not moved — it stays the recording
site for the seven transcript ledger keys that parity lock 15 drives. The
seven key constants move to core/grading/key_manifest.py so the relocated
evaluator imports no runner module beyond runner.models, and the five manifest
entries naming the deleted symbols are repointed at the shared function.

core/grading/transcript.py joins the runner subset (the runner's closure now
reaches it) and transcript_buckets.py takes its place among the
orchestrator-only exclusions.

tests/unit/grading/test_transcript.py is deleted: its evidence-set trio and
reason-string trio assert behaviour and a shape the shared evaluator does not
have, and every other test in it is already locked against the shared
evaluator in test_grading_correctness.py. The two concerns not covered there —
a role: system termination notice cannot satisfy a required phrase, and rows
are reproducible in declaration order — move into that module, beside a new
lock that a dumped config is refused.
