+++ docs/ARCHITECTURE.md
@@ -123,6 +123,7 @@ Component-level detail (CLI commands, the run-queue client, SecretManager, indiv
 | **Orchestrator + Core** | Loads run config, instantiates the adapter, builds the task list, manages the attempt queue, runs trials, aggregates grades and metrics, writes artifacts. | `tolokaforge/core` (`orchestrator.py`, `grading/`, `metrics.py`, `search/`) | [`docs/RUNNER.md`](RUNNER.md) |
 | **TrialRunner** | One instance per trial. Owns the agent–user message loop, calls the LLM for both agent and user simulator turns, and dispatches every tool call as a gRPC `ExecuteTool` RPC to the Runner service. | `tolokaforge/core/runner.py` | — |
 | **Trial event timeline** | Joins a trial's message trace and its tool-call record into one ordered event tuple, joined by `call_id`. A pure function in the grading path, so both grading substrates build the same timeline from the same trial. | `tolokaforge/core/grading/trace_timeline.py` | [`docs/GRADING.md`](GRADING.md#trial-event-timeline) |
+| **Transcript rules evaluator** | Scores a pack's `transcript_rules` block against the timeline, one sub-check per declared entry. Pure, and takes the validated `TranscriptRulesConfig` rather than a dump, so a rule cannot mean different things on the two substrates. The runner's `GradeTrial` scores the family through it; the core engine's own fold is being retired onto it (#685). | `tolokaforge/core/grading/transcript.py` | [`docs/GRADING.md`](GRADING.md#both-substrates-consume-it) |
 | **Adapter Layer** | Resolves a benchmark format into `(TaskConfig, tools, grading, environment, docker_stack_requirements, task_description)`. Built-in `native` plus plugins discovered via the `tolokaforge.adapters` entry-point group. | `tolokaforge/adapters` + `external_adapters/` | [`docs/ADAPTER_ARCHITECTURE.md`](ADAPTER_ARCHITECTURE.md) |
 | **LLM Layer** | Provider-agnostic completion API on top of LiteLLM. Per-provider presets carry capability policies for schema dialect, reasoning encoding, parameter handling, caching, tool-name discipline. Used by both agent and user-simulator turns. | `tolokaforge/core/llm` | [`docs/LLM_LAYER.md`](LLM_LAYER.md) |
 | **Runner Service (gRPC)** | The sole tool-execution path. Owns per-trial state, reconstructs tools from a `TaskDescription`, executes tool calls, runs grading. A single service — not split into separate "agent" and "executor" services. Auto-started by the orchestrator (`auto_start_services: true`). | `tolokaforge/runner` | [`docs/RUNNER.md`](RUNNER.md) · [`docs/GRPC_PROTOCOL.md`](GRPC_PROTOCOL.md) |
