+++ tolokaforge/core/_runner_subset.py
@@ -122,7 +122,7 @@ RUNNER_SUBSET_EXCLUDED_FILES: tuple[str, ...] = (
     "tolokaforge/core/grading/rubric_migration.py",
     "tolokaforge/core/grading/state_checks.py",
     "tolokaforge/core/grading/trace_replay.py",
-    "tolokaforge/core/grading/transcript.py",
+    "tolokaforge/core/grading/transcript_buckets.py",
     "tolokaforge/core/grading/unknown_keys.py",
     "tolokaforge/core/llm/fallback_client.py",
     "tolokaforge/tools/user_tools.py",
@@ -147,12 +147,13 @@ including any of these in the subset would drag those orchestrator-only
 surfaces along with them, or fail at import time inside the runner
 container. The remaining six (``core.grading.agreement``,
 ``core.grading.config_validation``, ``core.grading.migration_declaration``,
-``core.grading.transcript``, ``core.grading.unknown_keys``,
+``core.grading.transcript_buckets``, ``core.grading.unknown_keys``,
 ``core.llm.fallback_client``) have only shared-spine imports but are consumed
 exclusively by orchestrator-side code — the pre-run authoring gate and the
 rubric-to-trace-check migration both run on the host, before any trial is
-scheduled — and would ship as dead weight. The runner container's runtime
-closure reaches none of them.
+scheduled, and ``transcript_buckets`` is read only by the core engine's own
+fold — and would ship as dead weight. The runner container's runtime closure
+reaches none of them.
 """
