+++ tolokaforge/core/grading/key_manifest.py
@@ -162,8 +162,7 @@ class GradingKey:
             )
 
 
-_CORE_TRANSCRIPT_EVALUATOR = "tolokaforge.core.grading.transcript.TranscriptChecker.grade"
-_RUNNER_TRANSCRIPT_EVALUATOR = "tolokaforge.runner.grading.evaluate_transcript_rules"
+_TRANSCRIPT_EVALUATOR = "tolokaforge.core.grading.transcript.evaluate_transcript_rules"
 _CORE_HASH_EVALUATOR = "tolokaforge.core.grading.state_checks.StateChecker.check_hash"
 
 RUNNER_HASH_EVALUATOR = "tolokaforge.runner.service.RunnerServiceImpl._execute_hash_grading"
@@ -510,8 +509,8 @@ GRADING_KEYS: tuple[GradingKey, ...] = (
         enforcement=Enforcement.DIFFERENTIAL_CANONICAL,
         core_field="TranscriptRulesConfig.must_contain",
         runner_field="TranscriptRulesConfig.must_contain",
-        core_evaluator=_CORE_TRANSCRIPT_EVALUATOR,
-        runner_evaluator=_RUNNER_TRANSCRIPT_EVALUATOR,
+        core_evaluator=_TRANSCRIPT_EVALUATOR,
+        runner_evaluator=_TRANSCRIPT_EVALUATOR,
         reason=_TRANSCRIPT_PHRASE_REASON,
         tracking_issue=685,
     ),
@@ -522,8 +521,8 @@ GRADING_KEYS: tuple[GradingKey, ...] = (
         enforcement=Enforcement.DIFFERENTIAL_CANONICAL,
         core_field="TranscriptRulesConfig.disallow_regex",
         runner_field="TranscriptRulesConfig.disallow_regex",
-        core_evaluator=_CORE_TRANSCRIPT_EVALUATOR,
-        runner_evaluator=_RUNNER_TRANSCRIPT_EVALUATOR,
+        core_evaluator=_TRANSCRIPT_EVALUATOR,
+        runner_evaluator=_TRANSCRIPT_EVALUATOR,
         reason=_TRANSCRIPT_PHRASE_REASON,
         tracking_issue=685,
     ),
@@ -534,8 +533,8 @@ GRADING_KEYS: tuple[GradingKey, ...] = (
         enforcement=Enforcement.DIFFERENTIAL_CANONICAL,
         core_field="TranscriptRulesConfig.max_turns",
         runner_field="TranscriptRulesConfig.max_turns",
-        core_evaluator=_CORE_TRANSCRIPT_EVALUATOR,
-        runner_evaluator=_RUNNER_TRANSCRIPT_EVALUATOR,
+        core_evaluator=_TRANSCRIPT_EVALUATOR,
+        runner_evaluator=_TRANSCRIPT_EVALUATOR,
         reason=_TRANSCRIPT_AGGREGATION_REASON,
         tracking_issue=685,
     ),
@@ -546,8 +545,8 @@ GRADING_KEYS: tuple[GradingKey, ...] = (
         enforcement=Enforcement.DIFFERENTIAL_CANONICAL,
         core_field="TranscriptRulesConfig.min_assistant_turns",
         runner_field="TranscriptRulesConfig.min_assistant_turns",
-        core_evaluator=_CORE_TRANSCRIPT_EVALUATOR,
-        runner_evaluator=_RUNNER_TRANSCRIPT_EVALUATOR,
+        core_evaluator=_TRANSCRIPT_EVALUATOR,
+        runner_evaluator=_TRANSCRIPT_EVALUATOR,
     ),
     GradingKey(
         author_key="transcript_rules.required_actions",
@@ -557,7 +556,7 @@ GRADING_KEYS: tuple[GradingKey, ...] = (
         core_field="TranscriptRulesConfig.required_actions",
         runner_field="TranscriptRulesConfig.required_actions",
         core_evaluator="tolokaforge.core.evaluators.action_evaluator.ActionEvaluator.evaluate_actions",
-        runner_evaluator=_RUNNER_TRANSCRIPT_EVALUATOR,
+        runner_evaluator=_TRANSCRIPT_EVALUATOR,
         reason=_TRANSCRIPT_AGGREGATION_REASON,
         tracking_issue=685,
     ),
@@ -572,7 +571,7 @@ GRADING_KEYS: tuple[GradingKey, ...] = (
             "tolokaforge.core.evaluators.communicate_evaluator."
             "CommunicateEvaluator.evaluate_communication"
         ),
-        runner_evaluator=_RUNNER_TRANSCRIPT_EVALUATOR,
+        runner_evaluator=_TRANSCRIPT_EVALUATOR,
         reason=_TRANSCRIPT_AGGREGATION_REASON,
         tracking_issue=685,
     ),
@@ -583,10 +582,8 @@ GRADING_KEYS: tuple[GradingKey, ...] = (
         enforcement=Enforcement.DIFFERENTIAL_CANONICAL,
         core_field="TranscriptRulesConfig.tool_expectations",
         runner_field="TranscriptRulesConfig.tool_expectations",
-        core_evaluator=(
-            "tolokaforge.core.grading.transcript.TranscriptChecker.check_tool_expectations"
-        ),
-        runner_evaluator=_RUNNER_TRANSCRIPT_EVALUATOR,
+        core_evaluator=_TRANSCRIPT_EVALUATOR,
+        runner_evaluator=_TRANSCRIPT_EVALUATOR,
         reason=(
             "core folds both tool lists into one of four averaged buckets and ignores call "
             "status; the runner scores one sub-check per declared tool and requires a "
@@ -739,6 +736,14 @@ NO_TIMELINE_EVENTS_SKIP = KeyAccountingRecord(
 UNBOUND_BINDING_SKIP = KeyAccountingRecord(
     outcome=KeyAccounting.SKIPPED, detail="the binding yielded no assignment"
 )
+
+MUST_CONTAIN_KEY = checked_author_key("transcript_rules.must_contain")
+DISALLOW_REGEX_KEY = checked_author_key("transcript_rules.disallow_regex")
+MAX_TURNS_KEY = checked_author_key("transcript_rules.max_turns")
+MIN_ASSISTANT_TURNS_KEY = checked_author_key("transcript_rules.min_assistant_turns")
+TOOL_EXPECTATIONS_KEY = checked_author_key("transcript_rules.tool_expectations")
+REQUIRED_ACTIONS_KEY = checked_author_key("transcript_rules.required_actions")
+COMMUNICATE_INFO_KEY = checked_author_key("transcript_rules.communicate_info")
 
 TRACE_CONSTRAINTS_KEY = checked_author_key("trace_checks.constraints")
