+++ tolokaforge/core/grading/transcript.py
@@ -1,198 +1,483 @@
-"""Transcript-based grading rules"""
+"""Scoring a pack's transcript rules against the trial's event timeline.
+
+Substrate-neutral and pure — no services, no I/O — over the timeline both
+substrates build and the one ``TranscriptRulesConfig`` both substrates validate,
+so a rule reaches the same verdict whichever substrate grades the trial. The
+config is taken as the model, never as a dump: a dict seam admits a field
+spelling or a rule shape neither substrate agreed to, and admits it silently.
+
+:func:`evaluate_transcript_rules` serves two integration points — the runner's
+``GradeTrial`` and the core :class:`~tolokaforge.core.grading.combine.GradingEngine`
+— the way :func:`~tolokaforge.core.grading.trace_checks.evaluate_trace_checks`
+already does.
+
+The authored vocabulary is documented in ``docs/GRADING.md`` § "Transcript Rules".
+"""
 
 import re
+from collections.abc import Sequence
 
+from tolokaforge.core.grading.key_manifest import (
+    COMMUNICATE_INFO_KEY,
+    DISALLOW_REGEX_KEY,
+    EVALUATED,
+    MAX_TURNS_KEY,
+    MIN_ASSISTANT_TURNS_KEY,
+    MUST_CONTAIN_KEY,
+    REQUIRED_ACTIONS_KEY,
+    TOOL_EXPECTATIONS_KEY,
+)
 from tolokaforge.core.grading.trace_timeline import (
-    TraceEvent,
-    TraceEventKind,
+    AttemptedCall,
     TrialTimeline,
     assistant_texts,
     attempted_calls,
 )
-
-# What a phrase rule searches: the trial's own text. ``role: system`` harness
-# annotations are not timeline events, so a required phrase can never be
-# satisfied by a termination notice the harness wrote.
-_SEARCHABLE_KINDS = (
-    TraceEventKind.USER_MESSAGE,
-    TraceEventKind.ASSISTANT_MESSAGE,
-    TraceEventKind.TOOL_RESULT,
+from tolokaforge.runner.models import (
+    CommunicateInfo,
+    KeyAccountingRecord,
+    RequiredAction,
+    ToolExecutionStatus,
+    ToolExecutorIdentity,
+    TranscriptEvaluationResult,
+    TranscriptRuleResult,
+    TranscriptRulesConfig,
 )
 
 
-def _event_text(event: TraceEvent) -> str:
-    """A turn's text, or a tool result's recorded output."""
-    if event.kind is TraceEventKind.TOOL_RESULT:
-        return event.result or ""
-    return event.text or ""
+def evaluate_transcript_rules(
+    timeline: TrialTimeline,
+    rules: TranscriptRulesConfig,
+) -> TranscriptEvaluationResult:
+    """
+    Evaluate the author-facing ``TranscriptRulesConfig`` against the trial's
+    event timeline.
 
+    ``rules`` is the validated model task authors write in ``grading.yaml``,
+    decomposed into one independent sub-check per field entry, each of which
+    produces a visible ``TranscriptRuleResult`` in ``details``:
 
-def _searchable_text(timeline: TrialTimeline) -> str:
-    """Everything said or returned during the trial, as one string."""
-    return " ".join(
-        _event_text(event) for event in timeline.events if event.kind in _SEARCHABLE_KINDS
+    - ``must_contain`` (list[str]): each string must appear (case-insensitive
+      substring) in some assistant message → one sub-check per string.
+    - ``disallow_regex`` (list[str]): none of the regexes may match any
+      assistant message → one sub-check per regex.
+    - ``max_turns`` (int | None): the number of assistant turns must be within
+      the limit → one sub-check when set.
+    - ``min_assistant_turns`` (int | None): the trial must have produced at least
+      this many assistant turns. A **gate on the whole component**, not a
+      sub-check: an unmet floor appends one failing row and forces ``score`` to
+      ``0.0``, and a met floor appends no row so it never dilutes the fraction.
+    - ``tool_expectations`` ({required_tools, disallowed_tools}): one sub-check
+      per declared tool. A required tool must have been called *successfully*; a
+      disallowed tool must not have run at **any** status, because an attempted
+      forbidden call is itself the violation. Both fail on a timeline that carries
+      no records for a call the message view declared — whether it ran is then
+      unknown, and a "did not run" reading would pass every forbidden call.
+    - ``required_actions`` (list[RequiredAction]): each declared tool call must
+      appear on the timeline, matched by ``name`` + ``requestor`` and
+      by the argument subset named in ``compare_args`` (``None`` = compare all
+      declared args, ``[]`` = compare none) → one sub-check per action.
+    - ``communicate_info`` (list[CommunicateInfo]): each ``required`` info
+      string must appear in an assistant message → one sub-check per required
+      entry (non-required entries are advisory and not scored).
+
+    Scoring: ``score`` is the fraction of sub-checks that passed — unless a
+    declared ``min_assistant_turns`` floor is unmet, in which case it is ``0.0``
+    whatever the other sub-checks did. ``passed`` is True iff every sub-check
+    passed. The component score feeds
+    ``combine_grade_components`` where ``pass_threshold`` is applied, so a
+    fraction (rather than all-or-nothing) lets authors set partial-credit
+    thresholds (e.g. ``pass_threshold: 0.75``). A config that produces no sub-check
+    rows is a no-op pass with ``score=1.0`` — no fields set, or a met floor as the
+    only declared rule, leaves nothing that could have been violated.
+
+    Unknown / missing data is surfaced as a FAILING sub-check, never silently
+    passed (AGENTS.md: surface failures explicitly).
+
+    ``accounted_keys`` names the author keys this call decomposed, for the
+    runner's runtime ledger. The core engine has no ledger and discards it: the
+    accounting is a record of what one ``GradeTrial`` RPC did, not a property of
+    the score.
+
+    Args:
+        timeline: The trial's event timeline, built by
+            :func:`~tolokaforge.core.grading.trace_timeline.build_trial_timeline`
+        rules: The validated ``TranscriptRulesConfig``
+
+    Returns:
+        TranscriptEvaluationResult with passed, score, per-sub-check details, and
+        the author keys this call decomposed
+
+    Raises:
+        TypeError: ``rules`` is not a ``TranscriptRulesConfig``. The evaluator is
+            the one place the two substrates meet, and a dump reopens the untyped
+            seam this signature exists to close — a caller handing one over is
+            refused rather than silently reading a shape it cannot validate.
+    """
+    if not isinstance(rules, TranscriptRulesConfig):
+        raise TypeError(
+            "evaluate_transcript_rules takes a validated TranscriptRulesConfig, not "
+            f"{type(rules).__name__}. Pass the model both substrates already hold"
+        )
+
+    details: list[TranscriptRuleResult] = []
+    accounted_keys: dict[str, KeyAccountingRecord] = {}
+
+    tool_expectations = rules.tool_expectations
+    required_tools = tool_expectations.required_tools if tool_expectations else []
+    disallowed_tools = tool_expectations.disallowed_tools if tool_expectations else []
+
+    assistant_messages = assistant_texts(timeline)
+    calls = attempted_calls(timeline)
+    records_present = timeline.records_present
+
+    unmet_floor: TranscriptRuleResult | None = None
+    if rules.min_assistant_turns is not None:
+        accounted_keys[MIN_ASSISTANT_TURNS_KEY] = EVALUATED
+        unmet_floor = _unmet_activity_floor(rules.min_assistant_turns, len(assistant_messages))
+        if unmet_floor is not None:
+            details.append(unmet_floor)
+
+    if rules.must_contain:
+        accounted_keys[MUST_CONTAIN_KEY] = EVALUATED
+        for text in rules.must_contain:
+            details.append(_check_must_contain(text, assistant_messages))
+
+    if rules.disallow_regex:
+        accounted_keys[DISALLOW_REGEX_KEY] = EVALUATED
+        for pattern in rules.disallow_regex:
+            details.append(_check_disallow_regex(pattern, assistant_messages))
+
+    if rules.max_turns is not None:
+        accounted_keys[MAX_TURNS_KEY] = EVALUATED
+        details.append(_check_max_turns(rules.max_turns, len(assistant_messages)))
+
+    if required_tools or disallowed_tools:
+        accounted_keys[TOOL_EXPECTATIONS_KEY] = EVALUATED
+        for tool_name in required_tools:
+            details.append(_check_required_tool(tool_name, calls, records_present=records_present))
+        for tool_name in disallowed_tools:
+            details.append(
+                _check_disallowed_tool(tool_name, calls, records_present=records_present)
+            )
+
+    if rules.required_actions:
+        accounted_keys[REQUIRED_ACTIONS_KEY] = EVALUATED
+        for action in rules.required_actions:
+            details.append(_check_required_action(action, calls, records_present=records_present))
+
+    if rules.communicate_info:
+        accounted_keys[COMMUNICATE_INFO_KEY] = EVALUATED
+        for info in rules.communicate_info:
+            check = _check_communicate_info(info, assistant_messages)
+            if check is not None:
+                details.append(check)
+
+    if not details:
+        # No sub-check rows — nothing can be violated. Either nothing was declared,
+        # or a met activity floor was the only declaration and emits no row.
+        return TranscriptEvaluationResult(
+            passed=True, score=1.0, details=[], accounted_keys=accounted_keys
+        )
+
+    passed_count = sum(1 for d in details if d.passed)
+    total_count = len(details)
+    # An unmet activity floor vetoes the whole component: a trial that did not act
+    # makes the other sub-checks' verdicts evidence of nothing, and a fraction
+    # would let any pass_threshold at or below it swallow the violation.
+    score = 0.0 if unmet_floor is not None else passed_count / total_count
+    all_passed = passed_count == total_count
+
+    return TranscriptEvaluationResult(
+        passed=all_passed, score=score, details=details, accounted_keys=accounted_keys
     )
 
 
-class TranscriptChecker:
-    """Check conversation transcript against rules"""
+# Map the RequiredAction.requestor vocabulary ("assistant"/"user", the
+# author-facing role names) onto the recorded executor identity. They name the
+# same actor.
+_REQUESTOR_TO_EXECUTOR: dict[str, ToolExecutorIdentity] = {
+    "assistant": ToolExecutorIdentity.AGENT,
+    "user": ToolExecutorIdentity.USER,
+}
 
-    def check_must_contain(
-        self, timeline: TrialTimeline, phrases: list[str]
-    ) -> tuple[float, list[str]]:
-        """Check if transcript contains required phrases"""
-        if not phrases:
-            return 1.0, []
 
-        full_transcript = _searchable_text(timeline)
+_UNRECORDED = (
+    "the trial carries no tool-call record, so nothing knows whether the calls it "
+    "declared ran. A timeline rebuilt from a bundle's trajectory.yaml alone is in "
+    "that state — the record is the bundle's tool_log.yaml sidecar"
+)
 
-        found = 0
-        reasons = []
 
-        for phrase in phrases:
-            if phrase in full_transcript:
-                found += 1
-            else:
-                reasons.append(f"Missing required phrase: '{phrase}'")
+def _declared(calls: Sequence[AttemptedCall], tool_name: str) -> list[AttemptedCall]:
+    """Every call to ``tool_name`` the agent asked for, whether or not it ran."""
+    return [call for call in calls if call.tool_name == tool_name]
 
-        score = found / len(phrases) if phrases else 1.0
-        return score, reasons
 
-    def check_disallowed_regex(
-        self, timeline: TrialTimeline, patterns: list[str]
-    ) -> tuple[float, list[str]]:
-        """Check if transcript contains disallowed patterns"""
-        if not patterns:
-            return 1.0, []
+def _executed(calls: Sequence[AttemptedCall], tool_name: str) -> list[AttemptedCall]:
+    """The calls to ``tool_name`` that actually ran, so carry an outcome.
 
-        full_transcript = _searchable_text(timeline)
+    Separates ran from did-not-run only while the timeline carries records.
+    Without them every call reads as un-run, so callers ask
+    :func:`_outcome_unknown` first instead of reading absent evidence as a fact.
+    """
+    return [call for call in _declared(calls, tool_name) if call.status is not None]
 
-        violations = []
-        for pattern in patterns:
-            matches = re.findall(pattern, full_transcript, re.IGNORECASE)
-            if matches:
-                violations.append(f"Disallowed pattern '{pattern}' found: {matches[:3]}")
 
-        score = 0.0 if violations else 1.0
-        return score, violations
+def _outcome_unknown(
+    calls: Sequence[AttemptedCall], tool_name: str, *, records_present: bool
+) -> bool:
+    """Whether the timeline can say if the declared calls to ``tool_name`` ran.
 
-    def check_max_turns(self, timeline: TrialTimeline, max_turns: int | None) -> tuple[float, str]:
-        """Check if conversation stayed within turn limit"""
-        if max_turns is None:
-            return 1.0, ""
+    A record can only name a call the message view declared, so a tool the agent
+    never asked for never ran — knowable from the message view alone. Whatever it
+    did ask for is unknown the moment the record view is missing.
+    """
+    return not records_present and bool(_declared(calls, tool_name))
 
-        actual_turns = len(assistant_texts(timeline))
 
-        if actual_turns <= max_turns:
-            return 1.0, ""
-        else:
-            return 0.0, f"Exceeded max turns: {actual_turns} > {max_turns}"
+def _check_must_contain(text: str, assistant_messages: Sequence[str]) -> TranscriptRuleResult:
+    """Each required string must appear (case-insensitive) in some assistant message."""
+    needle = text.lower()
+    found = any(needle in content.lower() for content in assistant_messages)
+    return TranscriptRuleResult(
+        rule_type="must_contain",
+        rule={"must_contain": text},
+        passed=found,
+        message=(
+            f"Found required text {text!r} in an assistant message"
+            if found
+            else f"Required text {text!r} not found in any assistant message"
+        ),
+    )
 
-    def check_min_assistant_turns(
-        self, timeline: TrialTimeline, min_assistant_turns: int | None
-    ) -> tuple[float, str]:
-        """Check the agent produced at least the declared number of turns.
 
-        A "turn" is one assistant generation, the same counter ``max_turns``
-        bounds from above.
-        """
-        if min_assistant_turns is None:
-            return 1.0, ""
+def _check_disallow_regex(pattern: str, assistant_messages: Sequence[str]) -> TranscriptRuleResult:
+    """No assistant message may match the disallowed regex.
 
-        actual_turns = len(assistant_texts(timeline))
-
-        if actual_turns >= min_assistant_turns:
-            return 1.0, ""
-        return 0.0, (
-            f"Assistant turn count {actual_turns} below min_assistant_turns "
-            f"of {min_assistant_turns}"
+    An invalid regex is an author error — surface it as a FAIL rather than
+    silently treating it as 'no match'.
+    """
+    try:
+        compiled = re.compile(pattern, re.IGNORECASE)
+    except re.error as exc:
+        return TranscriptRuleResult(
+            rule_type="disallow_regex",
+            rule={"disallow_regex": pattern},
+            passed=False,
+            message=f"Invalid disallow_regex {pattern!r}: {exc}",
         )
 
-    def check_tool_expectations(
-        self,
-        timeline: TrialTimeline,
-        required_tools: list[str] | None,
-        disallowed_tools: list[str] | None,
-    ) -> tuple[float, list[str]]:
-        """Check tool usage expectations.
+    for content in assistant_messages:
+        if compiled.search(content):
+            return TranscriptRuleResult(
+                rule_type="disallow_regex",
+                rule={"disallow_regex": pattern},
+                passed=False,
+                message=f"Disallowed pattern {pattern!r} matched an assistant message",
+            )
+    return TranscriptRuleResult(
+        rule_type="disallow_regex",
+        rule={"disallow_regex": pattern},
+        passed=True,
+        message=f"Disallowed pattern {pattern!r} did not match any assistant message",
+    )
 
-        A call with no record did not run — but only while the timeline carries
-        records at all. Without them nothing knows whether the declared calls ran,
-        so the check fails naming that rather than reading absent evidence as
-        "never used", which would pass every disallowed tool the trial called.
-        """
-        reasons = []
-        score = 1.0
 
-        calls = attempted_calls(timeline)
-        if not timeline.records_present and calls and (required_tools or disallowed_tools):
-            declared = ", ".join(sorted({call.tool_name for call in calls}))
-            return 0.0, [
-                "Tool expectations unevaluatable: the trial carries no tool-call record, "
-                f"so whether it ran the calls it declared ({declared}) is unknown"
-            ]
+def _check_max_turns(max_turns: int, turn_count: int) -> TranscriptRuleResult:
+    """Assistant turn count must be within the limit.
 
-        tools_used = {call.tool_name for call in calls if call.status is not None}
+    A "turn" is one assistant generation. This counts the agent's responses
+    rather than user messages so the limit caps the agent's activity, which is
+    what authors intend to bound.
+    """
+    within = turn_count <= max_turns
+    return TranscriptRuleResult(
+        rule_type="max_turns",
+        rule={"max_turns": max_turns},
+        passed=within,
+        message=(
+            f"Assistant turn count {turn_count} within limit of {max_turns}"
+            if within
+            else f"Assistant turn count {turn_count} exceeds limit of {max_turns}"
+        ),
+    )
 
-        if required_tools:
-            missing = set(required_tools) - tools_used
-            if missing:
-                score *= 0.5
-                # Sort so the reason text is reproducible across runs — Python set
-                # repr ordering is not stable, which made these reasons drift
-                # spuriously across machines / processes.
-                reasons.append(f"Missing required tools: {', '.join(sorted(missing))}")
 
-        if disallowed_tools:
-            violations = tools_used & set(disallowed_tools)
-            if violations:
-                score = 0.0
-                reasons.append(f"Used disallowed tools: {', '.join(sorted(violations))}")
+def _unmet_activity_floor(min_assistant_turns: int, turn_count: int) -> TranscriptRuleResult | None:
+    """The failing row for an unmet activity floor, or ``None`` when it is met.
 
-        return score, reasons
+    A met floor emits no row at all. The floor is a gate on the whole component,
+    so a passing row would dilute the fraction the other sub-checks produce.
+    """
+    if turn_count >= min_assistant_turns:
+        return None
+    return TranscriptRuleResult(
+        rule_type="min_assistant_turns",
+        rule={"min_assistant_turns": min_assistant_turns},
+        passed=False,
+        message=(
+            f"Assistant turn count {turn_count} below min_assistant_turns of {min_assistant_turns}"
+        ),
+    )
 
-    def grade(
-        self,
-        timeline: TrialTimeline,
-        must_contain: list[str] | None = None,
-        disallow_regex: list[str] | None = None,
-        max_turns: int | None = None,
-        min_assistant_turns: int | None = None,
-        required_tools: list[str] | None = None,
-        disallowed_tools: list[str] | None = None,
-    ) -> tuple[float, str]:
-        """
-        Grade transcript with all rules
 
-        Returns:
-            (score 0-1, reasons)
-        """
-        all_reasons = []
+def _check_required_tool(
+    tool_name: str, calls: Sequence[AttemptedCall], *, records_present: bool
+) -> TranscriptRuleResult:
+    """A required tool must have been called successfully at least once.
 
-        # Check each rule
-        contain_score, contain_reasons = self.check_must_contain(timeline, must_contain or [])
-        all_reasons.extend(contain_reasons)
-
-        regex_score, regex_reasons = self.check_disallowed_regex(timeline, disallow_regex or [])
-        all_reasons.extend(regex_reasons)
-
-        turns_score, turns_reason = self.check_max_turns(timeline, max_turns)
-        if turns_reason:
-            all_reasons.append(turns_reason)
-
-        tools_score, tools_reasons = self.check_tool_expectations(
-            timeline, required_tools, disallowed_tools
+    Same "a failed call did not happen" rule ``_check_required_action`` applies:
+    an errored call did not accomplish the work the author required.
+    """
+    if _outcome_unknown(calls, tool_name, records_present=records_present):
+        return TranscriptRuleResult(
+            rule_type="required_tool",
+            rule={"required_tool": tool_name},
+            passed=False,
+            message=f"Required tool {tool_name!r}: {_UNRECORDED}",
         )
-        all_reasons.extend(tools_reasons)
+    called = any(call.status is ToolExecutionStatus.SUCCESS for call in _executed(calls, tool_name))
+    return TranscriptRuleResult(
+        rule_type="required_tool",
+        rule={"required_tool": tool_name},
+        passed=called,
+        message=(
+            f"Required tool {tool_name!r} was called successfully"
+            if called
+            else f"Required tool {tool_name!r} was never called successfully"
+        ),
+    )
 
-        floor_score, floor_reason = self.check_min_assistant_turns(timeline, min_assistant_turns)
-        if floor_reason:
-            all_reasons.append(floor_reason)
 
-        # The activity floor gates the average rather than joining it: a fifth
-        # zero bucket scores 0.8, which is the default ``pass_threshold``, so a
-        # declared floor would be unable to fail a trial.
-        scores = [contain_score, regex_score, turns_score, tools_score]
-        final_score = sum(scores) / len(scores) * floor_score
+def _check_disallowed_tool(
+    tool_name: str, calls: Sequence[AttemptedCall], *, records_present: bool
+) -> TranscriptRuleResult:
+    """A disallowed tool must not have run, at any status.
 
-        return final_score, "; ".join(all_reasons) if all_reasons else "All checks passed"
+    Status-insensitive on purpose: attempting a forbidden call is the violation,
+    so an errored attempt fails the check just like a successful one. A call the
+    agent declared on a terminating turn never reached the substrate and is not
+    counted; a ``trace_checks`` matcher over the call event is where that intent is
+    nameable. That exclusion needs the record view, so a declared call the timeline holds no
+    records for fails the check instead of reading as one that never ran.
+    """
+    if _outcome_unknown(calls, tool_name, records_present=records_present):
+        return TranscriptRuleResult(
+            rule_type="disallowed_tool",
+            rule={"disallowed_tool": tool_name},
+            passed=False,
+            message=f"Disallowed tool {tool_name!r}: {_UNRECORDED}",
+        )
+    offending = _executed(calls, tool_name)
+    if not offending:
+        message = f"Disallowed tool {tool_name!r} was never called"
+    else:
+        statuses = ", ".join(sorted({call.status.value for call in offending if call.status}))
+        message = (
+            f"Disallowed tool {tool_name!r} was called {len(offending)} "
+            f"time(s) (statuses: {statuses})"
+        )
+    return TranscriptRuleResult(
+        rule_type="disallowed_tool",
+        rule={"disallowed_tool": tool_name},
+        passed=not offending,
+        message=message,
+    )
+
+
+def _check_required_action(
+    action: RequiredAction, calls: Sequence[AttemptedCall], *, records_present: bool
+) -> TranscriptRuleResult:
+    """A declared tool call must appear on the timeline.
+
+    Matching:
+    - ``name`` must match the called tool exactly;
+    - ``requestor`` ("assistant"/"user") must match the call's ``executor``
+      ("agent"/"user");
+    - the call must have ``status == "success"`` (a failed call did not happen);
+    - arguments named by ``compare_args`` must match the declared values.
+      ``compare_args is None`` compares every declared argument; ``[]`` compares
+      none (presence of the tool call is enough).
+
+    An empty ``name`` is authorable — the field is required but not bounded — and
+    names no tool, so it is refused rather than matched against every call.
+    """
+    declared = action.model_dump()
+    expected_executor = _REQUESTOR_TO_EXECUTOR.get(action.requestor)
+
+    if action.compare_args is None:
+        keys_to_compare = list(action.arguments.keys())
+    else:
+        keys_to_compare = list(action.compare_args)
+
+    label = (
+        f"required_action {action.action_id!r} "
+        f"(tool={action.name!r}, requestor={action.requestor!r})"
+    )
+
+    if not action.name:
+        return TranscriptRuleResult(
+            rule_type="required_action",
+            rule=declared,
+            passed=False,
+            message=f"{label}: no name declared",
+        )
+
+    if _outcome_unknown(calls, action.name, records_present=records_present):
+        return TranscriptRuleResult(
+            rule_type="required_action",
+            rule=declared,
+            passed=False,
+            message=f"{label}: {_UNRECORDED}",
+        )
+
+    for call in calls:
+        if call.tool_name != action.name:
+            continue
+        if expected_executor is not None and call.executor is not expected_executor:
+            continue
+        if call.status is not ToolExecutionStatus.SUCCESS:
+            continue
+        if all(call.arguments.get(k) == action.arguments.get(k) for k in keys_to_compare):
+            return TranscriptRuleResult(
+                rule_type="required_action",
+                rule=declared,
+                passed=True,
+                message=f"{label}: matched a successful tool call",
+            )
+
+    return TranscriptRuleResult(
+        rule_type="required_action",
+        rule=declared,
+        passed=False,
+        message=(
+            f"{label}: no matching successful tool call found"
+            + (f" with args {keys_to_compare}" if keys_to_compare else "")
+        ),
+    )
+
+
+def _check_communicate_info(
+    info: CommunicateInfo, assistant_messages: Sequence[str]
+) -> TranscriptRuleResult | None:
+    """A required info string must appear in an assistant message.
+
+    Returns ``None`` (no sub-check, not scored) for non-required info — those
+    are advisory hints, not gating requirements.
+    """
+    if not info.required:
+        return None
+
+    needle = info.info.lower()
+    found = any(needle in content.lower() for content in assistant_messages)
+    return TranscriptRuleResult(
+        rule_type="communicate_info",
+        rule=info.model_dump(),
+        passed=found,
+        message=(
+            f"Communicated required info {info.info!r}"
+            if found
+            else f"Required info {info.info!r} was not communicated to the user"
+        ),
+    )
