+++ tolokaforge/runner/grading.py
@@ -3,7 +3,6 @@ Grading Helper Functions for Runner Service
 
 This module provides helper functions for the GradeTrial RPC implementation:
 - compute_state_diff: Compute human-readable diff between two stable states
-- evaluate_transcript_rules: Evaluate transcript rules against the trial's event timeline
 - combine_grade_components: Combine component scores into final grade
 - compose_runner_trial_verdict: Apply the judge and trace gates around that fold, so an
   offline recomputation reaches the runner's verdict without repeating either gate
@@ -13,8 +12,6 @@ See docs/GRPC_PROTOCOL.md for grading algorithm specification.
 
 import glob
 import logging
-import re
-from collections.abc import Sequence
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
@@ -45,31 +42,10 @@ from tolokaforge.core.grading.state_composition import (
     compose_state_checks_score,
     inert_hash_weight_reason,
 )
-from tolokaforge.core.grading.trace_timeline import (
-    AttemptedCall,
-    TrialTimeline,
-    assistant_texts,
-    attempted_calls,
-)
 from tolokaforge.core.hash import canonical_number
-from tolokaforge.runner.grading_ledger import (
-    COMMUNICATE_INFO_KEY,
-    DISALLOW_REGEX_KEY,
-    EVALUATED,
-    MAX_TURNS_KEY,
-    MIN_ASSISTANT_TURNS_KEY,
-    MUST_CONTAIN_KEY,
-    REQUIRED_ACTIONS_KEY,
-    TOOL_EXPECTATIONS_KEY,
-)
 from tolokaforge.runner.models import (
-    KeyAccountingRecord,
     StateDiff,
     TableDiff,
-    ToolExecutionStatus,
-    ToolExecutorIdentity,
-    TranscriptEvaluationResult,
-    TranscriptRuleResult,
 )
 
 logger = logging.getLogger(__name__)
@@ -247,435 +223,6 @@ def _get_field_diffs(expected: dict[str, Any], actual: dict[str, Any]) -> list[d
             diffs.append({"field": field, "expected": exp_val, "actual": act_val})
 
     return diffs
-
-
-def evaluate_transcript_rules(
-    timeline: TrialTimeline,
-    rules: dict[str, Any],
-) -> TranscriptEvaluationResult:
-    """
-    Evaluate the author-facing ``TranscriptRulesConfig`` against the trial's
-    event timeline.
-
-    ``rules`` is a single ``TranscriptRulesConfig.model_dump()`` dict — the
-    schema task authors actually write in ``grading.yaml``. It is decomposed
-    into one independent sub-check per field entry, each of which produces a
-    visible ``TranscriptRuleResult`` in ``details``:
-
-    - ``must_contain`` (list[str]): each string must appear (case-insensitive
-      substring) in some assistant message → one sub-check per string.
-    - ``disallow_regex`` (list[str]): none of the regexes may match any
-      assistant message → one sub-check per regex.
-    - ``max_turns`` (int | None): the number of assistant turns must be within
-      the limit → one sub-check when set.
-    - ``min_assistant_turns`` (int | None): the trial must have produced at least
-      this many assistant turns. A **gate on the whole component**, not a
-      sub-check: an unmet floor appends one failing row and forces ``score`` to
-      ``0.0``, and a met floor appends no row so it never dilutes the fraction.
-    - ``tool_expectations`` ({required_tools, disallowed_tools}): one sub-check
-      per declared tool. A required tool must have been called *successfully*; a
-      disallowed tool must not have run at **any** status, because an attempted
-      forbidden call is itself the violation. Both fail on a timeline that carries
-      no records for a call the message view declared — whether it ran is then
-      unknown, and a "did not run" reading would pass every forbidden call.
-    - ``required_actions`` (list[RequiredAction]): each declared tool call must
-      appear on the timeline, matched by ``name`` + ``requestor`` and
-      by the argument subset named in ``compare_args`` (``None`` = compare all
-      declared args, ``[]`` = compare none) → one sub-check per action.
-    - ``communicate_info`` (list[{info, required}]): each ``required`` info
-      string must appear in an assistant message → one sub-check per required
-      entry (non-required entries are advisory and not scored).
-
-    Scoring: ``score`` is the fraction of sub-checks that passed — unless a
-    declared ``min_assistant_turns`` floor is unmet, in which case it is ``0.0``
-    whatever the other sub-checks did. ``passed`` is True iff every sub-check
-    passed. The component score feeds
-    ``combine_grade_components`` where ``pass_threshold`` is applied, so a
-    fraction (rather than all-or-nothing) lets authors set partial-credit
-    thresholds (e.g. ``pass_threshold: 0.75``). A config that produces no sub-check
-    rows is a no-op pass with ``score=1.0`` — no fields set, or a met floor as the
-    only declared rule, leaves nothing that could have been violated.
-
-    Unknown / missing data is surfaced as a FAILING sub-check, never silently
-    passed (AGENTS.md: surface failures explicitly).
-
-    Args:
-        timeline: The trial's event timeline, built by
-            :func:`~tolokaforge.core.grading.trace_timeline.build_trial_timeline`
-        rules: A single ``TranscriptRulesConfig.model_dump()`` dict
-
-    Returns:
-        TranscriptEvaluationResult with passed, score, per-sub-check details, and
-        the author keys this call decomposed
-    """
-    details: list[TranscriptRuleResult] = []
-    accounted_keys: dict[str, KeyAccountingRecord] = {}
-
-    must_contain: list[str] = rules.get("must_contain", []) or []
-    disallow_regex: list[str] = rules.get("disallow_regex", []) or []
-    max_turns: int | None = rules.get("max_turns")
-    min_assistant_turns: int | None = rules.get("min_assistant_turns")
-    tool_expectations: dict[str, Any] = rules.get("tool_expectations") or {}
-    required_tools: list[str] = tool_expectations.get("required_tools") or []
-    disallowed_tools: list[str] = tool_expectations.get("disallowed_tools") or []
-    required_actions: list[dict[str, Any]] = rules.get("required_actions", []) or []
-    communicate_info: list[dict[str, Any]] = rules.get("communicate_info", []) or []
-
-    assistant_messages = assistant_texts(timeline)
-    calls = attempted_calls(timeline)
-    records_present = timeline.records_present
-
-    unmet_floor: TranscriptRuleResult | None = None
-    if min_assistant_turns is not None:
-        accounted_keys[MIN_ASSISTANT_TURNS_KEY] = EVALUATED
-        unmet_floor = _unmet_activity_floor(min_assistant_turns, len(assistant_messages))
-        if unmet_floor is not None:
-            details.append(unmet_floor)
-
-    if must_contain:
-        accounted_keys[MUST_CONTAIN_KEY] = EVALUATED
-        for text in must_contain:
-            details.append(_check_must_contain(text, assistant_messages))
-
-    if disallow_regex:
-        accounted_keys[DISALLOW_REGEX_KEY] = EVALUATED
-        for pattern in disallow_regex:
-            details.append(_check_disallow_regex(pattern, assistant_messages))
-
-    if max_turns is not None:
-        accounted_keys[MAX_TURNS_KEY] = EVALUATED
-        details.append(_check_max_turns(max_turns, len(assistant_messages)))
-
-    if required_tools or disallowed_tools:
-        accounted_keys[TOOL_EXPECTATIONS_KEY] = EVALUATED
-        for tool_name in required_tools:
-            details.append(_check_required_tool(tool_name, calls, records_present=records_present))
-        for tool_name in disallowed_tools:
-            details.append(
-                _check_disallowed_tool(tool_name, calls, records_present=records_present)
-            )
-
-    if required_actions:
-        accounted_keys[REQUIRED_ACTIONS_KEY] = EVALUATED
-        for action in required_actions:
-            details.append(_check_required_action(action, calls, records_present=records_present))
-
-    if communicate_info:
-        accounted_keys[COMMUNICATE_INFO_KEY] = EVALUATED
-        for info in communicate_info:
-            check = _check_communicate_info(info, assistant_messages)
-            if check is not None:
-                details.append(check)
-
-    if not details:
-        # No sub-check rows — nothing can be violated. Either nothing was declared,
-        # or a met activity floor was the only declaration and emits no row.
-        return TranscriptEvaluationResult(
-            passed=True, score=1.0, details=[], accounted_keys=accounted_keys
-        )
-
-    passed_count = sum(1 for d in details if d.passed)
-    total_count = len(details)
-    # An unmet activity floor vetoes the whole component: a trial that did not act
-    # makes the other sub-checks' verdicts evidence of nothing, and a fraction
-    # would let any pass_threshold at or below it swallow the violation.
-    score = 0.0 if unmet_floor is not None else passed_count / total_count
-    all_passed = passed_count == total_count
-
-    return TranscriptEvaluationResult(
-        passed=all_passed, score=score, details=details, accounted_keys=accounted_keys
-    )
-
-
-# Map the RequiredAction.requestor vocabulary ("assistant"/"user", the
-# author-facing role names) onto the recorded executor identity. They name the
-# same actor.
-_REQUESTOR_TO_EXECUTOR: dict[str, ToolExecutorIdentity] = {
-    "assistant": ToolExecutorIdentity.AGENT,
-    "user": ToolExecutorIdentity.USER,
-}
-
-
-_UNRECORDED = (
-    "the trial carries no tool-call record, so nothing knows whether the calls it "
-    "declared ran. A timeline rebuilt from a bundle's trajectory.yaml alone is in "
-    "that state — the record is the bundle's tool_log.yaml sidecar"
-)
-
-
-def _declared(calls: Sequence[AttemptedCall], tool_name: str) -> list[AttemptedCall]:
-    """Every call to ``tool_name`` the agent asked for, whether or not it ran."""
-    return [call for call in calls if call.tool_name == tool_name]
-
-
-def _executed(calls: Sequence[AttemptedCall], tool_name: str) -> list[AttemptedCall]:
-    """The calls to ``tool_name`` that actually ran, so carry an outcome.
-
-    Separates ran from did-not-run only while the timeline carries records.
-    Without them every call reads as un-run, so callers ask
-    :func:`_outcome_unknown` first instead of reading absent evidence as a fact.
-    """
-    return [call for call in _declared(calls, tool_name) if call.status is not None]
-
-
-def _outcome_unknown(
-    calls: Sequence[AttemptedCall], tool_name: str, *, records_present: bool
-) -> bool:
-    """Whether the timeline can say if the declared calls to ``tool_name`` ran.
-
-    A record can only name a call the message view declared, so a tool the agent
-    never asked for never ran — knowable from the message view alone. Whatever it
-    did ask for is unknown the moment the record view is missing.
-    """
-    return not records_present and bool(_declared(calls, tool_name))
-
-
-def _check_must_contain(text: str, assistant_messages: Sequence[str]) -> TranscriptRuleResult:
-    """Each required string must appear (case-insensitive) in some assistant message."""
-    needle = text.lower()
-    found = any(needle in content.lower() for content in assistant_messages)
-    return TranscriptRuleResult(
-        rule_type="must_contain",
-        rule={"must_contain": text},
-        passed=found,
-        message=(
-            f"Found required text {text!r} in an assistant message"
-            if found
-            else f"Required text {text!r} not found in any assistant message"
-        ),
-    )
-
-
-def _check_disallow_regex(pattern: str, assistant_messages: Sequence[str]) -> TranscriptRuleResult:
-    """No assistant message may match the disallowed regex.
-
-    An invalid regex is an author error — surface it as a FAIL rather than
-    silently treating it as 'no match'.
-    """
-    try:
-        compiled = re.compile(pattern, re.IGNORECASE)
-    except re.error as exc:
-        return TranscriptRuleResult(
-            rule_type="disallow_regex",
-            rule={"disallow_regex": pattern},
-            passed=False,
-            message=f"Invalid disallow_regex {pattern!r}: {exc}",
-        )
-
-    for content in assistant_messages:
-        if compiled.search(content):
-            return TranscriptRuleResult(
-                rule_type="disallow_regex",
-                rule={"disallow_regex": pattern},
-                passed=False,
-                message=f"Disallowed pattern {pattern!r} matched an assistant message",
-            )
-    return TranscriptRuleResult(
-        rule_type="disallow_regex",
-        rule={"disallow_regex": pattern},
-        passed=True,
-        message=f"Disallowed pattern {pattern!r} did not match any assistant message",
-    )
-
-
-def _check_max_turns(max_turns: int, turn_count: int) -> TranscriptRuleResult:
-    """Assistant turn count must be within the limit.
-
-    A "turn" is one assistant generation. This counts the agent's responses
-    rather than user messages so the limit caps the agent's activity, which is
-    what authors intend to bound.
-    """
-    within = turn_count <= max_turns
-    return TranscriptRuleResult(
-        rule_type="max_turns",
-        rule={"max_turns": max_turns},
-        passed=within,
-        message=(
-            f"Assistant turn count {turn_count} within limit of {max_turns}"
-            if within
-            else f"Assistant turn count {turn_count} exceeds limit of {max_turns}"
-        ),
-    )
-
-
-def _unmet_activity_floor(min_assistant_turns: int, turn_count: int) -> TranscriptRuleResult | None:
-    """The failing row for an unmet activity floor, or ``None`` when it is met.
-
-    A met floor emits no row at all. The floor is a gate on the whole component,
-    so a passing row would dilute the fraction the other sub-checks produce.
-    """
-    if turn_count >= min_assistant_turns:
-        return None
-    return TranscriptRuleResult(
-        rule_type="min_assistant_turns",
-        rule={"min_assistant_turns": min_assistant_turns},
-        passed=False,
-        message=(
-            f"Assistant turn count {turn_count} below min_assistant_turns of {min_assistant_turns}"
-        ),
-    )
-
-
-def _check_required_tool(
-    tool_name: str, calls: Sequence[AttemptedCall], *, records_present: bool
-) -> TranscriptRuleResult:
-    """A required tool must have been called successfully at least once.
-
-    Same "a failed call did not happen" rule ``_check_required_action`` applies:
-    an errored call did not accomplish the work the author required.
-    """
-    if _outcome_unknown(calls, tool_name, records_present=records_present):
-        return TranscriptRuleResult(
-            rule_type="required_tool",
-            rule={"required_tool": tool_name},
-            passed=False,
-            message=f"Required tool {tool_name!r}: {_UNRECORDED}",
-        )
-    called = any(call.status is ToolExecutionStatus.SUCCESS for call in _executed(calls, tool_name))
-    return TranscriptRuleResult(
-        rule_type="required_tool",
-        rule={"required_tool": tool_name},
-        passed=called,
-        message=(
-            f"Required tool {tool_name!r} was called successfully"
-            if called
-            else f"Required tool {tool_name!r} was never called successfully"
-        ),
-    )
-
-
-def _check_disallowed_tool(
-    tool_name: str, calls: Sequence[AttemptedCall], *, records_present: bool
-) -> TranscriptRuleResult:
-    """A disallowed tool must not have run, at any status.
-
-    Status-insensitive on purpose: attempting a forbidden call is the violation,
-    so an errored attempt fails the check just like a successful one. A call the
-    agent declared on a terminating turn never reached the substrate and is not
-    counted; a ``trace_checks`` matcher over the call event is where that intent is
-    nameable. That exclusion needs the record view, so a declared call the timeline holds no
-    records for fails the check instead of reading as one that never ran.
-    """
-    if _outcome_unknown(calls, tool_name, records_present=records_present):
-        return TranscriptRuleResult(
-            rule_type="disallowed_tool",
-            rule={"disallowed_tool": tool_name},
-            passed=False,
-            message=f"Disallowed tool {tool_name!r}: {_UNRECORDED}",
-        )
-    offending = _executed(calls, tool_name)
-    if not offending:
-        message = f"Disallowed tool {tool_name!r} was never called"
-    else:
-        statuses = ", ".join(sorted({call.status.value for call in offending if call.status}))
-        message = (
-            f"Disallowed tool {tool_name!r} was called {len(offending)} "
-            f"time(s) (statuses: {statuses})"
-        )
-    return TranscriptRuleResult(
-        rule_type="disallowed_tool",
-        rule={"disallowed_tool": tool_name},
-        passed=not offending,
-        message=message,
-    )
-
-
-def _check_required_action(
-    action: dict[str, Any], calls: Sequence[AttemptedCall], *, records_present: bool
-) -> TranscriptRuleResult:
-    """A declared tool call must appear on the timeline.
-
-    Matching:
-    - ``name`` must match the called tool exactly;
-    - ``requestor`` ("assistant"/"user") must match the call's ``executor``
-      ("agent"/"user");
-    - the call must have ``status == "success"`` (a failed call did not happen);
-    - arguments named by ``compare_args`` must match the declared values.
-      ``compare_args is None`` compares every declared argument; ``[]`` compares
-      none (presence of the tool call is enough).
-    """
-    action_id = action.get("action_id", "")
-    tool_name = action.get("name", "")
-    requestor = action.get("requestor", "")
-    declared_args = action.get("arguments", {}) or {}
-    compare_args = action.get("compare_args")
-
-    expected_executor = _REQUESTOR_TO_EXECUTOR.get(requestor)
-
-    if compare_args is None:
-        keys_to_compare = list(declared_args.keys())
-    else:
-        keys_to_compare = list(compare_args)
-
-    label = f"required_action {action_id!r} (tool={tool_name!r}, requestor={requestor!r})"
-
-    if not tool_name:
-        return TranscriptRuleResult(
-            rule_type="required_action",
-            rule=action,
-            passed=False,
-            message=f"{label}: no name declared",
-        )
-
-    if _outcome_unknown(calls, tool_name, records_present=records_present):
-        return TranscriptRuleResult(
-            rule_type="required_action",
-            rule=action,
-            passed=False,
-            message=f"{label}: {_UNRECORDED}",
-        )
-
-    for call in calls:
-        if call.tool_name != tool_name:
-            continue
-        if expected_executor is not None and call.executor is not expected_executor:
-            continue
-        if call.status is not ToolExecutionStatus.SUCCESS:
-            continue
-        if all(call.arguments.get(k) == declared_args.get(k) for k in keys_to_compare):
-            return TranscriptRuleResult(
-                rule_type="required_action",
-                rule=action,
-                passed=True,
-                message=f"{label}: matched a successful tool call",
-            )
-
-    return TranscriptRuleResult(
-        rule_type="required_action",
-        rule=action,
-        passed=False,
-        message=(
-            f"{label}: no matching successful tool call found"
-            + (f" with args {keys_to_compare}" if keys_to_compare else "")
-        ),
-    )
-
-
-def _check_communicate_info(
-    info: dict[str, Any], assistant_messages: Sequence[str]
-) -> TranscriptRuleResult | None:
-    """A required info string must appear in an assistant message.
-
-    Returns ``None`` (no sub-check, not scored) for non-required info — those
-    are advisory hints, not gating requirements.
-    """
-    text = info.get("info", "")
-    required = info.get("required", True)
-    if not required:
-        return None
-
-    needle = text.lower()
-    found = any(needle in content.lower() for content in assistant_messages)
-    return TranscriptRuleResult(
-        rule_type="communicate_info",
-        rule=info,
-        passed=found,
-        message=(
-            f"Communicated required info {text!r}"
-            if found
-            else f"Required info {text!r} was not communicated to the user"
-        ),
-    )
 
 
 def evaluate_jsonpath_file_checks(
